requirements.txt
python-telegram-bot==13.1
requests
BeautifulSoup4
sympy
mtranslate
